feat: navigate tool
