+++ ee/hogai/graph/root/tools/__init__.py
@@ -1,3 +1,4 @@
+from .navigate import NavigateTool, NavigateToolArgs
 from .read import ReadDataTool, ReadDataToolArgs
 from .read_taxonomy import ReadTaxonomyTool
 from .search import SearchTool, SearchToolArgs
@@ -11,4 +12,6 @@ __all__ = [
     "ReadDataToolArgs",
     "TodoWriteTool",
     "TodoWriteToolArgs",
+    "NavigateTool",
+    "NavigateToolArgs",
 ]
